requirements.txt
# Python 2.6 compatibility
# argparse==1.2.1
ipaddress==1.0.17
python-nmap==0.6.1
